requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
Pillow==10.4.0
python-dotenv==1.1.1
psycopg2-binary==2.9.11
dj-database-url==3.0.1